+++ packages/patternfly-4/react-core/src/components/Wizard/Wizard.tsx
@@ -7,7 +7,7 @@ import styles from '@patternfly/patternfly/components/Wizard/wizard.css';
 import { Backdrop } from '../Backdrop';
 import { Bullseye } from '../../layouts/Bullseye';
 import { WizardHeader } from './WizardHeader';
-import { WizardFooterInternal } from './WizardFooter';
+import { WizardFooterInternal } from './WizardFooterInternal';
 import { WizardToggle } from './WizardToggle';
 import { WizardNav } from './WizardNav';
 import { WizardNavItem } from './WizardNavItem';
@@ -60,7 +60,7 @@ export interface WizardProps {
   /** The wizard description */
   description?: string;
   /** Callback function to close the wizard */
-  onClose?(): void;
+  onClose?: () => void;
   /** Callback function when a step in the nav is clicked */
   onGoToStep?: WizardStepFunctionType;
   /** Additional classes spread to the Wizard */
@@ -76,7 +76,7 @@ export interface WizardProps {
   /** (Use to control the footer) Passing in a footer component lets you control the buttons yourself */
   footer?: React.ReactNode;
   /** (Unused if footer is controlled) Callback function to save at the end of the wizard, if not specified uses onClose */
-  onSave?(): void;
+  onSave?: () => void;
   /** (Unused if footer is controlled) Callback function after Next button is clicked */
   onNext?: WizardStepFunctionType;
   /** (Unused if footer is controlled) Callback function after Back button is clicked */
@@ -91,19 +91,19 @@ export interface WizardProps {
   ariaLabelCloseButton?: string;
 }
 
-export class Wizard extends React.Component<WizardProps> {
+interface WizardState {
+  currentStep: number;
+  isNavOpen: boolean;
+}
+
+export class Wizard extends React.Component<WizardProps, WizardState> {
   private static currentId = 0;
-  private static defaultProps = {
+  static defaultProps = {
     isOpen: false,
     isCompactNav: false,
     isFullHeight: false,
     isFullWidth: false,
-    width: null,
-    height: null,
     description: '',
-    onBack: null,
-    onNext: null,
-    onGoToStep: null,
     className: '',
     startAtStep: 1,
     nextButtonText: 'Next',
@@ -112,31 +112,34 @@ export class Wizard extends React.Component<WizardProps> {
     ariaLabelCloseButton: 'Close',
     ariaLabelNav: 'Steps',
     hasBodyPadding: true,
-    footer: null
+    onBack: null as WizardStepFunctionType,
+    onNext: null as WizardStepFunctionType,
+    onGoToStep: null as WizardStepFunctionType,
+    width: null as string,
+    height: null as string,
+    footer: null as React.ReactNode
   };
-
-  public container?: HTMLDivElement = undefined;
-  public titleId = `pf-wizard-title-0`;
-  public descriptionId = `pf-wizard-description-0`;
+  private container: HTMLDivElement;
+  private titleId: string;
+  private descriptionId: string;
 
   constructor(props: WizardProps) {
     super(props);
     const newId = Wizard.currentId++;
     this.titleId = `pf-wizard-title-${newId}`;
     this.descriptionId = `pf-wizard-description-${newId}`;
+    this.state = {
+      currentStep: this.props.startAtStep && Number.isInteger(this.props.startAtStep) ? this.props.startAtStep : 1,
+      isNavOpen: false
+    };
   }
-
-  public state = {
-    currentStep: this.props.startAtStep && Number.isInteger(this.props.startAtStep) ? this.props.startAtStep : 1,
-    isNavOpen: false
-  };
 
   private handleKeyClicks = (event: KeyboardEvent): void => {
     if (event.keyCode === KEY_CODES.ESCAPE_KEY) {
       if (this.state.isNavOpen) {
         this.setState({ isNavOpen: !this.state.isNavOpen })
       } else if (this.props.isOpen) {
-        this.props.onClose!();
+        this.props.onClose();
       }
     }
   };
@@ -160,7 +163,7 @@ export class Wizard extends React.Component<WizardProps> {
       if (onSave) {
         return onSave();
       }
-      return onClose!();
+      return onClose();
     } else {
       const newStep = currentStep + 1;
       this.setState({
@@ -266,8 +269,8 @@ export class Wizard extends React.Component<WizardProps> {
     // Set default Step values
     for (let i = 0; i < steps.length; i++) {
       if (steps[i].steps) {
-        for (let j = 0; j < steps[i].steps!.length; j++) {
-          steps[i].steps![j] = Object.assign({ canJumpTo: true }, steps[i].steps![j])
+        for (let j = 0; j < steps[i].steps.length; j++) {
+          steps[i].steps[j] = Object.assign({ canJumpTo: true }, steps[i].steps[j])
         }
       }
       steps[i] = Object.assign({ canJumpTo: true }, steps[i]);
@@ -275,7 +278,7 @@ export class Wizard extends React.Component<WizardProps> {
     return steps;
   }
 
-  public componentDidMount() {
+  componentDidMount() {
     if (this.container) {
       document.body.appendChild(this.container);
     }
@@ -283,7 +286,7 @@ export class Wizard extends React.Component<WizardProps> {
     document.addEventListener('keydown', this.handleKeyClicks, false);
   }
 
-  public componentWillUnmount() {
+  componentWillUnmount() {
     if (this.container) {
       document.body.removeChild(this.container);
     }
@@ -291,7 +294,7 @@ export class Wizard extends React.Component<WizardProps> {
     document.removeEventListener('keydown', this.handleKeyClicks, false);
   }
 
-  public render() {
+  render() {
     if (!canUseDOM) {
       return null;
     }
@@ -417,9 +420,34 @@ export class Wizard extends React.Component<WizardProps> {
           <Backdrop>
             <Bullseye>
               <WizardContextProvider value={context}>
-                <div {...rest} className={css(styles.wizard, isCompactNav && 'pf-m-compact-nav', activeStep.isFinishedStep && 'pf-m-finished', setFullWidth && styles.modifiers.fullWidth, setFullHeight && styles.modifiers.fullHeight, className)} role="dialog" aria-modal="true" aria-labelledby={this.titleId} aria-describedby={description ? this.descriptionId : undefined}>
-                  <WizardHeader titleId={this.titleId} descriptionId={this.descriptionId} onClose={onClose} title={title} description={description as string} ariaLabelCloseButton={ariaLabelCloseButton as string} />
-                  <WizardToggle isNavOpen={this.state.isNavOpen} onNavToggle={isNavOpen => this.setState({ isNavOpen })} nav={nav} steps={steps} activeStep={activeStep} hasBodyPadding={hasBodyPadding as boolean}>
+                <div {...rest}
+                  className={css(
+                    styles.wizard,
+                    isCompactNav && 'pf-m-compact-nav',
+                    activeStep.isFinishedStep && 'pf-m-finished',
+                    setFullWidth && styles.modifiers.fullWidth,
+                    setFullHeight && styles.modifiers.fullHeight,
+                    className)}
+                  role="dialog"
+                  aria-modal="true"
+                  aria-labelledby={this.titleId}
+                  aria-describedby={description ? this.descriptionId : undefined}
+                >
+                  <WizardHeader
+                    titleId={this.titleId}
+                    descriptionId={this.descriptionId}
+                    onClose={onClose}
+                    title={title} 
+                    description={description}
+                    ariaLabelCloseButton={ariaLabelCloseButton} />
+                  <WizardToggle
+                    isNavOpen={this.state.isNavOpen}
+                    onNavToggle={isNavOpen => this.setState({ isNavOpen })}
+                    nav={nav}
+                    steps={steps}
+                    activeStep={activeStep}
+                    hasBodyPadding={hasBodyPadding}
+                  >
                     {footer || (
                       <WizardFooterInternal
                         onNext={this.onNext}
@@ -439,7 +467,7 @@ export class Wizard extends React.Component<WizardProps> {
             </Bullseye>
           </Backdrop>
         </FocusTrap>,
-        this.container!
+        this.container
       )
     );
   }