+++ packages/patternfly-4/react-docs/gatsby-node.js
@@ -31,6 +31,7 @@ exports.createPages = ({ graphql, actions }) => {
           title
           section
           fullscreen
+          typescript
         }
       }
     }
@@ -60,12 +61,13 @@ exports.createPages = ({ graphql, actions }) => {
       } else {
         // Normal templated component pages
         link = `/${section}/${componentName}/`.toLowerCase();
-        // console.log('adding page', link);
+        // console.log('adding page', link, node.frontmatter.typescript);
         actions.createPage({
           path: link,
           component: path.resolve('./src/templates/mdxTemplate.js'),
           context: {
             title: node.frontmatter.title,
+            typescript: node.frontmatter.typescript, // For a badge
             fileAbsolutePath: node.fileAbsolutePath, // Helps us get the markdown
             pathRegex: `/${folderName}\/.*/` // Helps us get the docgenned props
           }
@@ -85,7 +87,7 @@ exports.onCreateWebpackConfig = ({ stage, actions }) => {
         '@patternfly-safe/react-core': '@patternfly/react-core',
         '@patternfly-safe/react-icons': '@patternfly/react-icons',
         '@patternfly-safe/react-tokens': '@patternfly/react-tokens',
-        // Resolve imports in .mdx files to local files
+        // Resolve imports in .mdx files to local dist folders
         '@patternfly/react-charts': path.resolve(__dirname, '../react-charts'),
         '@patternfly/react-core': path.resolve(__dirname, '../react-core'),
         '@patternfly/react-icons': path.resolve(__dirname, '../../react-icons'),