+++ packages/patternfly-4/react-docs/plugins/gatsby-transformer-react-docgen-typescript/gatsby-node.js
@@ -1,5 +1,4 @@
-const fs = require('fs')
-const docgenJavascript = require('react-docgen')
+const reactDocgen = require('react-docgen');
 
 function isSource(node) {
   if (!node ||
@@ -52,17 +51,9 @@ async function onCreateNode({
   const sourceText = await loadNodeContent(node);
   let parsed = null;
   try {
-    if (isTSX(node)) {
-      // Load up the already transpiled file
-      const transpiledPath = node.absolutePath
-        .replace(/react-(.*)\/src/, 'react-$1/dist/esm')
-        .replace(/\.tsx?$/, '.js');
-      const jsText = fs.readFileSync(transpiledPath, 'utf8');
-      parsed = docgenJavascript.parse(jsText);
-    }
-    else if (isJSX(node)) {
-      parsed = docgenJavascript.parse(sourceText);
-    }
+    parsed = reactDocgen.parse(sourceText, null, null, {
+      filename: node.absolutePath
+    });
   } catch (err) {
     console.warn('No component found in', node.absolutePath);
   }
@@ -87,4 +78,4 @@ async function onCreateNode({
   }
 }
 
-exports.onCreateNode = onCreateNode
+exports.onCreateNode = onCreateNode;